refactor: move useVariantUrl to utils/variants for better organization and update Cart component import path for clarity
