+++ app/components/cart/Cart.tsx
@@ -13,7 +13,7 @@ import {
 import type { CartLineUpdateInput } from "@shopify/hydrogen/storefront-api-types";
 import clsx from "clsx";
 import type { CartApiQueryFragment } from "storefrontapi.generated";
-import { useVariantUrl } from "~/lib/variants";
+import { useVariantUrl } from "~/lib/utils/variants";
 import { cn } from "~/lib/utils";
 import { IconRemove } from "../Icon";
