+++ app/lib/utils/variants.ts
